Deploy: do not error out if sources were not found
When 'wult' project is installed via an OS package (e.g., RPM), the sources for
the drivers and helpers are not provided. Instead, all deployables are just
deployed by the RPM, so 'wult deploy' is not even supposed to be used.

For this reason, this patch changes 'is_deploy_needed()' to avoid
raising an error if the sources were not found.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>