Deploy: provide 'get_installed_helper_path()'
This patch is based on the observation that helpers path does not have to be
the same path where 'wult deploy' installs them. Indeed, if 'wult' is installed
via an OS package, helpers can installed anywhere.

This patch adds the 'get_installed_helper_path()' function which tries to
figure out path of the installed helper.

This patch is needed to support wult installation via an RPM package.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>