Deploy: split the 'Deploy()' class
The 'Deploy.Deploy()' class provided two pieces of functionality:
 1 Deploying installables.
 2 Verifying that all installables are deployed.

This works well for the 'pip' model, when wult is installed using 'pip' (or by
just cloning the sources) and 'wult deploy' installs everything.

But now we are trying to support the package model (e.g., RPM). In this case
wult and all the dependencies are installed from the OS package. The 'wult
deploy' command becomes almost useless. And when we verify the installables, we
do not have the sources, so we should not try comparing time-stamps.

In order to support the RPM model, the "deploy()' and 'check_deployment()'
methods need to be heavily modified. After the modifications, they become very
different and do not really fit one class.

Therefore, this patch splits the 'Deploy()' class into 'Deploy()' and
'DeployCheck()'. There is some duplication left, but it will be taken care of
in the next patches.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>