Deploy: check only the required installables
This patch is big, but here what it does at high level.

Before this patch, the 'DeployCheck()' class accepted information about all the
possible installables, then it would "adjust" the installables, removing those
that cannot satisfied, and then it would check if the rest of the installables
exists.

This logic is complex and not optimal, because we end up checking for more
things than we need, and we end up printing useless warnings. For example, if
someone runs wult with 'hrtimer' method, we check for drivers (even though they
are not needed for 'hrtimer), we warn about kernel not being new enough for a
driver.

This patch changes the logic. Now the caller of
'DeployCheck.__init__()' has to provide only the required installables in
'deploy_info', instead of providing all of them. This allows us removing a lot
of code from 'DeployCheck' class.

This patch adds a 'ToolsCommon.reduce_installables()' helper that reduces the
isntallables to only the ones required by the used measurement method.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>