wultrunner: improve accuracy via usage of hrtimer_expire_entry tracepoint
Use hrtimer_expire_entry tracepoint to capture the IRQ timestamp
earlier. While doing this, also change the ordering of the capture of
the performance counters to happen in the latest event (IRQ vs wakeup),
which will improve the latency values of the one of these that happens
later in time.

Signed-off-by: Tero Kristo <tero.kristo@linux.intel.com>