TurbostatParser: fix bad use of null-like value
This patch fixes a warning raised by coverity:
    (CID: 2653016) Bad use of null-like value

This warning was raised because it was possible that the variables
'pkgdata' and 'coredata' would not be initialised before being used in a
predicate of an if-statement. This patch fixes the warning by
initialising the variables as empty dictionaries outside of
if-statements.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>