wult: drivers: allow idle hook to re-run after early exit
If CPU idle runs before timer expire, currently it marks the timer as
not-armed. This leads in the current timer cycle to be completely
ignored for measurement purposes. This leads into large amount of events
to be completely ignored mostly in the POLL cstate measurement, where
each idle cycle only lasts for 100us. The timer event must be programmed
to this period so that the idle event will be captured by the tool.

Fix the issue by not tagging the timer disarmed, if it hasn't actually
fired yet. This allows the CPU to enter idle again, and to try to
capture the data during the next CPU idle round.

Signed-off-by: Tero Kristo <tero.kristo@linux.intel.com>