_ReportBase: restore diagrams in C-state tabs
C-state tabs were recently fixed so that they could be included in wult
reports. However, the fix did not include restoring diagrams in those
tabs. This patch restores the diagrams by reworking how plots are added
to tabs.

Before this patch, 'ReportBase' would attempt to add a base metric to a
plot if a definition existed for it in 'defs'. It did not check if the
results actually contained data for the base metric. This patch makes it
so that base metrics will only be used if results actually contain data
for them.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>