_ReportBase: separate 'SilentTime/LDist' tab generation
This tab should have a specific configuration of diagrams. This patch
separates the tab from the rest of the results tabs.

Specifically, this patch also introduces the concept of a tab
"configuration". Eventually all tabs may be based on some user-specified
configuration. Therefore it is possible that other tabs will switch to
being generated this way in future.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>