Devices: save restore operstate in bind/unbind
Do not save/restore the operational state __init__/close, because it is not
necessary in many use-cases, such as 'wult scan'. And it requires the 'ip'
tool. Instead, do it in bind/unbind, because this is really what changes the
operational state. After re-binding, it may be up or down depending the the OS,
and we just restore it.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>