misc: view_multiple_reports.py: add a warning to avoid GUI
By default 'view_multiple_reports.py' creates a 'Tkinter' GUI to prompt
the user to select which directory to serve.

If the user is on a system with no graphical interface, this will raise
an error. This patch adjusts the script so that it catches this error
and tells the user to use '--dir' to avoid it in future.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>