misc: view_multiple_reports.py: catch and handle tkinter import error
Even though 'tkinter' is part of the standard Python library, on some
systems it is required to be provided by an OS-library. This means it is
not guaranteed to be installed. Hence, this patch changes
'view_multiple_reports.py' so that it warns the user when 'tkinter' is
required and suggests that they install 'tkinter' or use the '--dir'
option.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>