js: report-page: merge 'file upload' with CORS warning
This patch merges the 'file upload' button into the CORS warning so that
when the browser fails to retrieve report files, the solution is clearly
explained to the report viewer.

By combining the two elements, the whole warning is now packaged inside
a shoelace dialog element which is not closeable.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>