wultrunner: fix divide by zero error when ldist-min == ldist-max
As part of the logic to calculate launch distance, eBPF program uses
modulo operator to shrink the random range between the min,max values.
If the min,max are the same, this causes a divide by zero error terminating
the eBPF program. Fix the issue by adding +1 to max value if they are equal.

Signed-off-by: Tero Kristo <tero.kristo@linux.intel.com>