_Collector: fix 'discover()' not finding all failed collectors
'discover()' should check all collectors to see if they are available.
If collectors are fallible then it means that if/when they fail during
discovery, 'discover()' crashes and stops checking other collectors. To
prevent this behaviour, this patch updates 'configure()' to mark all
collectors as fallible if 'discovery' is 'True'.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>