_Collector: rename the 'discovery' argument
This is minor readability improvement: rename the 'discovery' argument to
'for_discovery', in order to make the code a bit more readable. This name makes
it a bit more clear that we are configuring/initializing for the discovery
process. Otherwise the 'discovery' argument makes an impression that the method
should do the discovery.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>