WultRunner: move statistics configuration out to the caller
Change 'WultRunner' API and instead of accepting the 'stconf' dictionaly,
accept a 'WultStatsCollect' object, which is supposed to be prepared for usage
by the caller. So all 'WultRunner' needs to do is to start using it.

This simplifies the 'WultRunner' class and makes it possible for the caller to
configure various statistics properties, such as device node paths, etc.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>