StatsCollect: add 'is_stcagent_needed()' method
Add a 'StatsCollect.is_stcagent_needed()' method that can be used to check
whether the 'stc-agent' program is required for collecting the enabled
statistics.

Start using it in a couple of places, and remove custom hacks that were
checking for 'sysinfo' being the only statistics. The new helper is more
generic.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>