pelicanconf: fix links breaking with custom 'PORT' during serve
Before this patch, the base URL for serving documentation was
'localhost:8000' with the port hard-coded. If a different 'PORT' value
was given, it would not change the base URL. This broke many internal
links in the documentation while serving. This patch removes the
hard-coded port value to fix that behaviour.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>