js: tab-group: make text in non-leaf nodes of the tab-tree clickable
Before this patch, text in non-leaf nodes couldn't be clicked to expand
tree sections. This is because the tree was in a mode which allows
non-leaf nodes to be selected. This patch changes the selection mode so
that only leaf nodes can be selected and text in non-leaf nodes can be
clicked to expand tree sections.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>