ToolsCommon: fix wult start --stats acpower,turbostat
It was broken by the recent rework patches, and failed with this
symptom:

wult: error: 'stc-agent' did not respond with 'OK' to the following command:
configure
Instead, the response was the following:
error: the 'configure' method of the acpower collector failed:
the 'acpower' statistics collector failed:
please, configure 'devnode' first
Check 'stc-agent' log file':

The reason is that we skip configuring the device node for the power
meter in there is no discovery. But we should do it in any case, and
this patch fixes the issue.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>