StatsCollect: fix a bug in '_normalize_stnames()'
Previously, '_normalize_stnames()' resolved 'stnames="all"' or
'stnames=None' to all possible aggregate and specific statistic names.

This patch changes this behaviour so that it resolves to all specific
available statistics. Importantly, if the out-of-band agent is not
initialised, it will only resolve to in-band statistics.

Fixes the following error breaking 'wult start' when no statistics are
specified:

wult: error: unknown statistic name 'ipmi-oob', the known names are:
  sysinfo, turbostat, ipmi-inband

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>