wult: rename the 'hrtimer' method
The main change this patch does is renaming the 'hrtimer' method into
'hrt-bpf'. This is done because we will likely have another eBPF-based
method soon.

This rename made things look inconsistently, so this patch includes
other renames to eliminate the inconsistencies.

1. wutrunner -> wult-hrt-helper
2. ndlrunner -> ndl-helper

And many variables in various files are renamed accordingly.

This required regenerating the pre-compiled eBPF code, which was also
done.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>