StatsCollect: remove 'is_stcagent_needed()' override
Previously, this method assumed that all specific statistics of an
aggregate statistic are enabled if the aggregate statistic is. Whether
an aggregate statistic has been used to enable specific statistics
or not, does not impact if a 'STCAgent' is needed. For this reason, this
patch removes the 'is_stcagent_needed()' override.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>