Add exercise-sut tool
The 'exercise-sut' tool can be used to run payload
command for various system configuration
permutations.

For example, the tool can be requested to run
'wult' with requestable C-states 'C1' and 'C6'
when C-state prewake feature is 'on' and 'off'.
These options would result running 'wult' 4 times
when deepest requestable C-state and C-state
prewake values are as follows:
1. C-State: C1, prewake: on
2. C-State: C1, prewake: off
3. C-state: C6, prewake: on
4. C-state: C6, prewake: off

Signed-off-by: Antti Laakso <antti.laakso@intel.com>