Deploy: append new category data retaining existing
If there are multiple data entries with the same category, the whole
category content gets currently overridden, only retaining the last one
under wulttools/_Wult.py. This is a problem if we want to install
multiple bpfhelpers, so fix this by appending the data to existing _cats
dict entry instead.

Signed-off-by: Tero Kristo <tero.kristo@linux.intel.com>