js: report-page: change 'toolname' in report info to 'title'
Before this patch, 'toolname' in report info was just used to determine
the title of the report in the format '{toolname} Report'.

This patch clarifies the interface by making the info attribute reflect
its effect on the generated report. It also gives the caller more
freedom when titling the report.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>