_ProgressLine: fix messy warnings
Before this patch, warnings could end up looking messy like this:

Datapoints: 791, max. latency: 50.83 us, rate: 395.21 datapoints/secwult: warning: blah

This is because the logger assumes that it alwasy starts from a newline, which
is not the case with our progress bar.

Fix this problem by setting the 'LOG.force_tty_newline_prefix' to 'True', which
will force the logger to prepend the next message it prints with a new line.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>