_MetricDTabBuilder: fix summary funcs not being added to summary table
Before this patch, if the value for a summary function was 0, it would
not be added to the summary table as 0 is a falsy value. This patch
fixes this behaviour by correcting the appropriate condition to check if
a value is 'None' rather than any falsy value.

As a result, this also fixes a bug where large report generation would
fail while generating the summary tables with the 'LDist' metric where
all values are 0. In this case, it would try to dump the summary table
when there were no summary functions for a metric which was added. This
resulted in the following error:

Generating WakeLatency tab.
Traceback (most recent call last):
  File "wult/statscollectlibs/htmlreport/_SummaryTable.py", line 196, in <genexpr>
    funcs = max((len(dct[metric]) for dct in self.smrytbl["funcs"].values()))
KeyError: 'Launch Distance'

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>