js: report-page: prevent error in browser when report has no intro table
Before this patch, if a report was generated without an intro table
(which is optional) then the browser complained about a raised error as
it tried to parse a non-existant intro table source file.

This patch modifies 'report-page.js' so that it will only try to parse
an intro table if one is provided.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>