Deploy: improve error message
If we build locally but intent to install to the remote host, we fail with a
message like this:

$ wult deploy --local-build --kernel-src '/lib/modules/blah' -H xyz
wult: error: kernel modules directory '/lib/modules/blah' does not exist on host 'xyz'

It is confusing because it is not obvious why the tool looks for
'/lib/modules/blah' on 'xyz', while it was asked to build locally
(--local-build).

This patch improves the error message. Now it looks more like this:

wult: error: kernel modules directory '/lib/modules/blah' does not exist on host 'xyz'
Even though you are building on local host, the result will have to be installed to 'xyz'.
For this reason you should have same kernel version (blah) installed on 'xyz'

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>