Deploy: rename '_DeployBase' to '_KernelHelper'
'_DeployBase' has become more of a helper class which contains a couple
of methods to help with kernel module paths and versions. This patch
renames the class to reflect this. In a following patch, the class will
also no longer be used as a base class so will not require the 'Base'
suffix.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>