Deploy: separate '_cats' and '_insts' initialisation from base class
In preparation for changing 'Deploy' and 'DeployCheck' to use a
different base class, this patch isolates some of the common logic so
that it can still be reused without defining twice. Importantly, this
patch also weakens the linking between the base class and the child
classes. This will facilitate the switch to a new base class in future.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>