Deploy: change 'Deploy' and 'DeployCheck' interface to 'KernelHelper'
This patch changes 'Deploy' and 'DeployCheck' from inheriting from
'KernelHelper' to a composite relationship where they store an instance
of 'KernelHelper' as a class property.

This is to facilitate the creation of new base classes for 'Deploy' and
'DeployCheck' which do not depend on any of the methods provided by
'KernelHelper'. The new base classes will be provided by
'statscollectlibs'.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>