_STCAgent: add 'collect_sysinfo_after()'
Before this patch, in the case where 'inbagent.stop(sysinfo=True)' was
followed by 'oobagent.stop()', the out-of-band agent would be stopped
significantly later than the in-band agent (sometimes >1m) while the
'sysinfo' files were copied to the 'outdir'.

This meant that the statistics being collected on 'oobagent' would run
for longer and indicate a jump in resource usage and the end of the
collection.

This patch adds the 'collect_sysinfo_after()' method to correct this
behaviour so that both agents can be stopped before using resources to
copy sysinfo files.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>