wult-tdt-helper: fix the total cycles / cc0 cycles calculation logic
There is a considerable gap between the snapshot capturing of the TSC
and MPERF counter values, which leads into incorrect calculation of
cc0 residency percentage especially with very short idle durations.
Fix the issue by capturing the TSC and MPERF values explicitly always
during the snapshot capturing, and modify the logic to reverse order
when exiting idle.

Also, the number of perf events is passed to the eBPF program by the
userspace now, so that we always only try to access the available
performance counters during the snapshot capture.

Signed-off-by: Tero Kristo <tero.kristo@linux.intel.com>