wult-hrt-helper: Add support for absolute timers
During startup of the BPF program, check for the presence of BPF
absolute timer support. If it is available, use this one. Otherwise
fall back to the relative timer support.

Absolute timer has the benefit of being more accurate for latency
calculations, as we know the exact time when the timer is going to
expire; the BPF program directly programs the (absolute) expire time.
Relative timer on the other hand is started with an offset to the
current time, but there is no way to query back the exact expire
time that was actually programmed.

Signed-off-by: Tero Kristo <tero.kristo@linux.intel.com>