DFSummary: calc_col_smry: define behaviour for 'nan' values
Before this patch, if a summary function was calculated as a 'nan'
value, the function returned '{}, None'. This behaviour was not
documented and was dissimilar to the normal returned value. This patch
clarifies the case when a summary function is calculated as a 'nan'
value by setting the value to 'None' and continuing with calculating
other summary functions.

This patch also fixes the following error in 'wult report' when it is used to
generate a report for a dataset with a small amount of datapoints for a given
metric:

File "statscollectlibs/htmlreport/tabs/_DTabBuilder.py", line 54, in add_smrytbl
    self._smrytbl.add_smry_func(rep, mdef["title"], fname, smry_dict[fname])
TypeError: tuple indices must be integers or slices, not str

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>