DeployHelpersBase: rename 'helpername'
We have the 'helpername' variable used a lot in "*RawDataProvider" modules, and
using the same variable name in "Deploy*" modules is a bit confusing. Let's
rename it.

I did not come up with a name better than 'what'. But I think it is fine,
because we use 'what' in other places for similar purposes.

Se rename 'helpername' to 'what'. Also improve the values we pass in order to
have better deployment messages.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>