Add a temporary module (for deploy rework)
The deployment code is currenly somewhat messy and I am trying to rework it.
The goal is to simplify it and make it more logical. For example, wult's
'Deploy' should not inherit from stat-collect's deploy.

Doing this in small increments is very hard, so I'll try to do it in large
steps. Will start with 'stats-collect deploy'.

This patch copies 'statscollectlibs.Deploy' to 'wultlibs' as
'DeployTemporary', so 'wult deploy' will not break while I am reworking
'stats-collect deploy'. The temporary module will be removed later.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>