js: report-page: generate unique tab IDs to prevent unopenable tabs
Before this patch, the 'tab-panel' component used the names of tabs as
IDs. This caused a conflict when tabs have the same name (for example in
turbostat totals and measured CPU tabs).

This patch changes the 'report-page' to generate unique tab IDs when it
reads 'tabs.json' from the report files to avoid these conflicts.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>