_SummaryTable: fix 'generate()' crashing with 'N/A' values in cells
'add_smry_func()' supports creating a summary table with "N/A" values
when a value is unavailable for a given summary function. When more than
one result contains a "N/A" value, 'generate()' attempts to generate
hovertext based on the two values. At this point it crashes. This patch
fixes this behaviour so that hover text is not generated for any "N/A"
values.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>