_MetricTabBuilder: smry_tbl: add smry func with only one value
Before this patch, a summary function would only be added to a summary
table if every result has a value for the function.

'SummaryTable' was recently changed to allow 'N/A' values. This removes
the requirement for every result to have a value for each summary
function.

This patch changes '_MetricTabBuilder' so that summary tables in metric
tabs will include all summary functions with data available. This means
that rather than excluding any functions because one or more results do
not have a value for that summary function, all the data available will
be included.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>