wult-tdt-helper: avoid timing hazard with extremely short ldist
It is possible that the timer can expire during the execution of the
bpf_timer_start if the relative launch time is extremely short. In this
case, the timer_armed flag will remain 'true' indefinitely effectively
stopping it from sending any trace events to userspace.

Fix this problem by moving the timer_armed flag handling to happen
before actually arming the timer.

Signed-off-by: Tero Kristo <tero.kristo@linux.intel.com>