js: diagram: remove blank space while the diagram loads
Before this patch, while the diagram is loading the report shows a
loading indicator and an empty space which the diagram will fill once it
has loaded in.

This extra space looks strange and might make the user think something
is missing from the report. This patch fixes that behaviour by giving
the reserved space for the diagram a height of 0. Once the diagram is
loaded, the height is restored to the desired setting.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>