_STCAgent: do not remove logs in case of errors
In case of an error related to 'stc-agent', wult exits wit an error message like
this:

Check 'stc-agent' log file on host 'bdwup0'':
/tmp/stc-agent-3eBEua/logs/stc-agent-bdwup0.log.txt

However, the log file is not available, because it is removed upon exit. This
patch fixes the problem - now the 'stc-agent' output directory is preserved in
case of 'stc-agent' errors.

Also fix the double single quotes issue in the message while on it.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>