stc-agent: generalize property setting
Turn the 'do_set_collector_property()' function into the '_set_obj_property()'
method. The new method does not assume that the object it works with is a
collector object, so it can work with any other "compatible" object.

This will allow to have agent properties, and use '_set_obj_property()' for
setting agent properties in upcoming changes.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>