_SpecStatsCollect: save paths to labels and statistics
Save path to the labels file in 'stats-info.yml'. The path is saved for every
collected statistic, and it is relative to the 'stats' sub-directory.

Save statistics file/directory path too, so that it would be enough to read
'stats-info.yml' to find all statistics, without scanning the "stats"
sub-directory.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>