_ACPowerTabBuilder: fix 'stats-collect report' crashing on 'inf' values
Previously, 'ACPowerTabBuilder' was fixed to handle 'inf' values in raw
AC Power statististics files. The fix did not catch the case where the
removed datapoint is then included in scatter plot data reduction.

This patch fixes that edge case by resetting the index on the dataframe
in question. This also prevents any further issues which could have
arised because of inconsistent indexing.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>