_ReportBase: only include results with stats in 'stats_paths'
The 'statscollectlibs' 'HTMLReport.generate_report()' method accepts
'stats_paths' which should provide the paths to the statistics
directories for each result.

Before this patch, 'ReportBase' included results with no statistics in
the 'stats_paths' argument passed to 'statscollectlibs' 'HTMLReport'.

After this patch, 'stats_paths' only contains results with statistics to
avoid any undefined behaviour in 'statscollectlibs' 'HTMLReport'.

This will facilitate future development on 'statscollectlibs' 'HTMLReport'
so that it does not need to maintain this undocumented support.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>