_Plot: create_hover_template: fix hover text containing template text
Before this patch, if a given 'pandas.DataFrame' column was not
available for one of the datapoints (but not all) then that datapoint
would show some of the hover template text instead of data e.g.
'customdata[12]'.

The desired behaviour is for a datapoint to not show anything if data is
not available for that datapoint. This patch enforces that behaviour by
changing 'create_hover_template()' to generate a template for each
datapoint.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>