wultlibs: cleanup and improve 'isatty()' usage
There are few places where we use 'isatty()' to determine whether we are
dealing with a TTY or not. The difference is that a TTY "understands" special
characters, such as color codes and carriege return.

The problem with using 'isatty()', however, is that it does not honor wult's
and other tools "--force-color" option. This option is probably mis-named, and
it really means '--force-tty' - treat the statnard output and error streams as
a TTY, even if they are not.

This patch fixes the situation and changes all the places where we rely on
'isatty()' to use '_LOG.colored'. This way, if someone runs:

wult --force-color start > outfile.tx

wult will send the progress bar updates to 'outfile.txt'.

Renaming '--force-color' to '--force-tty' is a separate effort.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>