wult-tdt-helper: Discard ping event if event ringbuffer is not empty
While measuring CPU idle poll state, it is necessary to send the 'ping'
message to the CPU to wake it up from idle, otherwise the CPU will only
wakeup at next tick timer event. This is already taken care of, but in
cases where there is huge amount of data being sent, like if user has
configured to use extremely short launch distances for the tool, several
messages may pile up into the ring buffer. If there is already data in
the ringbuffer, attempting to send a new 'ping' message will not have
the desired effect of waking up the CPU, instead just buffering the
message, and skewing the wakeup latency for current measurement.
To avoid the problem, before sending a 'ping' message check if the
ringbuffer is empty and we can expect the CPU to really be woken up,
otherwise discard the current measurement completely.

Signed-off-by: Tero Kristo <tero.kristo@linux.intel.com>