stc-agent: have collectors stop at the same time
This patch "reverts" the changes done in:

664d2b21 stc-agent: fix falure in 'save()'

We wait for proceses in 'save()' for a good reason, eveon though this is hacky.
We first want to signall all of them in 'end()' without waiting, so that they
all start exiting. And then later we check if they exited in 'save()'.

Add a piece of commentary explaining this, move waiting for the process back to
'save()'.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>