_ReportBase: fix bug where results with no stats crashed 'wult report'
If statistics have been loaded successfully, 'res.stats_res' will be
populated. Otherwise, either there was no statistics to process or there
was an issue loading them.

This patch changes '_ReportBase' to use the 'stats_res' when determining
which results should be included in statistics tabs. As a result of this
change, it fixes a bug where 'wult' report generation would crash if a
result contained an empty or malformed 'stats-collect' result.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>