_TabBuilderBase: fix a bug causing some turbostat tabs not to generate
This patch fixes a bug where in the case that there are no common
metrics between two or more results for one or more container tabs, then
the turbostat tab fails to generate.

The patch fixes this by adjusting 'build_ctab()' to handle if a tab
hierarchy contains an empty container tab.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>