_BatchConfig: refactor to support tool-specific arguments
Today wult and ndl are aspecial, and we need to supply 'devid' to them. But
there are other special tools we want to support, with their own special
arguments. For example, 'stats-collect' needs the program to run.

Make 'devid' support less explicite, to make it easier to add other special
arguments later.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>