_BatchConfig: make 'cpu' argument "more optional"
Some special tools (wult and ndl) require the CPU number, but other special
tools (stats-collect and benchmark) don't. There are some methods that
explicitly take CPU number, but depending on the tool, do not use it. So CPU
number is optional.

This patch makes it even "more" optional by moving it into 'kwargs'. This way
if a class needs it, it will be found in kwargs. This way allows for passing
more arguments like 'cpu', which some tools need and some do not need.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>