wult-tdt-helper: replace the TSC calibration by direct tap to kernel
Instead of attempting to calibrate the TSC value from the relative
timebase to the absolute one, directly read the updated raw TSC value
in the eBPF program from the perf_event structure. We still need to
trigger a read from the eBPF program via the perf subsystem though,
because this causes the execution of the kernel function which updates
the raw value in the perf_event structure. This approach gets rid of any
calibration errors from the eBPF read TSC value.

Signed-off-by: Tero Kristo <tero.kristo@linux.intel.com>