_ReportBase: resolve outdir in 'wulttools'
This is in preparation to save the report generation log in the output
directory. That requires the output directory to be determined at the
'wulttools' level so that it can be passed to the logger.

Therefore, this patch moves the resolution of report directories inside
of results out of '_ReportBase' to facilitate the saving of report
generation logs.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>