RORawResult: only warn user if 'stats_dir' exists and 'stats_res' fails
If a result does not contain any statistics directory, then there is no
need to warn the user that statistics couldn't be loaded for the result.

This patch changes it so that the user will only be warned if the
statistics directory exists and an error occurs while building the
statistics result object.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>