_ReportBase: move '_copy_dir()' to 'statscollectlibs' 'HTMLReport'
This patch moves the '_copy_dir()' helper function to 'statscollectlibs'
so that it can be used by callers of 'HTMLReport'. This will allow the
callers to more easily add to report output directories.

Also, bump the version of 'stats-collect' required by 'wult' so that
'wult' uses the version which provides the 'copy_dir()' function.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>