_ReportBase: overwrite tool name and tool version in HTML Reports
The 'HTMLReport' class now displays the tool name and version used to
generate the report. This patch uses the 'toolname' and 'toolver'
parameters to override the tool details with the relevent name and
version for each tool in the 'wult' project.

This patch increments the 'stats-collect' requirement to use the new
'toolname' and 'toolver' options available in v1.0.14.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>